+++ mycodo/mycodo_flask/forms/forms_input.py
@@ -37,22 +37,23 @@ class InputAdd(FlaskForm):
     list_inputs_sorted = generate_form_input_list(dict_inputs)
 
     for each_input in list_inputs_sorted:
-        if 'interfaces' not in dict_inputs[each_input]:
-            choices_inputs.append(
-                ('{inp},'.format(inp=each_input),
-                 '{manuf}: {name}: {meas}'.format(
+        value = '{inp},'.format(inp=each_input)
+        name = '{manuf}: {name}'.format(
                      manuf=dict_inputs[each_input]['input_manufacturer'],
-                     name=dict_inputs[each_input]['input_name'],
-                     meas=dict_inputs[each_input]['measurements_name'])))
-        else:
+                     name=dict_inputs[each_input]['input_name'])
+
+        if 'input_library' in dict_inputs[each_input]:
+            name += ' ({lib})'.format(lib=dict_inputs[each_input]['input_library'])
+
+        name += ': {meas}'.format(meas=dict_inputs[each_input]['measurements_name'])
+
+        if 'interfaces' in dict_inputs[each_input]:
             for each_interface in dict_inputs[each_input]['interfaces']:
-                choices_inputs.append(
-                    ('{inp},{int}'.format(inp=each_input, int=each_interface),
-                     '{manuf}: {name}: {meas} ({int})'.format(
-                         manuf=dict_inputs[each_input]['input_manufacturer'],
-                         name=dict_inputs[each_input]['input_name'],
-                         meas=dict_inputs[each_input]['measurements_name'],
-                         int=each_interface)))
+                tmp_value = '{val},{int}'.format(val=value, int=each_interface)
+                tmp_name = '{name} ({int})'.format(name=name, int=each_interface)
+                choices_inputs.append((tmp_value, tmp_name))
+        else:
+            choices_inputs.append((value, name))
 
     input_type = SelectField(
         choices=choices_inputs,